+++ src/libs/blueprint/conduit_blueprint_mesh.cpp
@@ -2445,17 +2445,25 @@ mesh::generate_index(const Node &mesh,
     index_out.reset();
 
     index_out["state/number_of_domains"] = number_of_domains;
-    
-    // check if the input mesh has state/cycle state/time
-    // if so, add those to the index
-    if(mesh.has_path("state/cycle"))
-    {
-        index_out["state/cycle"].set(mesh["state/cycle"]);
-    }
 
-    if(mesh.has_path("state/time"))
+
+    if(mesh.has_child("state"))
     {
-        index_out["state/time"].set(mesh["state/time"]);
+        // check if the input mesh has state/cycle state/time
+        // if so, add those to the index
+        if(mesh.has_path("state/cycle"))
+        {
+            index_out["state/cycle"].set(mesh["state/cycle"]);
+        }
+
+        if(mesh.has_path("state/time"))
+        {
+            index_out["state/time"].set(mesh["state/time"]);
+        }
+        // state may contain other important stuff, like 
+        // the domain_id, so we need a way to read it
+        // from the index
+        index_out["state/path"] = join_path(ref_path, "state");
     }
 
     NodeConstIterator itr = mesh["coordsets"].children();